Refactor main into separate inventoryapp package
- Separate routes into their own handlers, each being a reciever function
on InventoryApp.
- Remove LogWrap as it was no longer needed due to wanting to pass errors
up from the Initialize methods into the main() method.

Missing unit tests to be added later.

https://github.com/edgexfoundry-holding/rfid-llrp-inventory-service/pull/18#discussion_r595535830
