+++ internal/inventory/app/app_test.go
@@ -1,0 +1,6 @@
+//
+// Copyright (C) 2021 Intel Corporation
+//
+// SPDX-License-Identifier: Apache-2.0
+
+package inventoryapp
